Created test project for the model.core project, and started GEF model.
The modeler.core.test project is a plain maven jar project, with no
src/main/java folder, and tests in src/test/java.

The idea is that this project will contain unit test for model.core,
which do not require an OSGi runtime startup, or are limited by the
packages exported by the manifest.

The new project has been used to TDD a Node class that will form part of
the model used by the GEF editor.